+++ geoip_website_redirect/__manifest__.py
@@ -20,10 +20,10 @@
 #
 #############################################################################
 {
-    'name': 'Website GeoIP Based Language and Currency',
+    'name': 'Website Geolocation Based Language and Currency',
     'version': '14.0.1.0.0',
     'category': 'eCommerce, Productivity',
-    'summary': """Website GeoIP Based Language and Currency Application
+    'summary': """Website Geolocation Based Language and Currency Application
      for Odoo 14""",
     'description': """The module helps you to map currency and language
      based on Customer's IP address""",