+++ geoip_website_redirect/doc/RELEASE_NOTES.md
@@ -4,5 +4,5 @@
 #### Version 14.0.1.0.0
 #### ADD
 
-- Initial commit for Website GeoIP Based Language and Currency
+- Initial commit for Website Geolocation Based Language and Currency